requirements.txt
flask 
pandas
numpy
scikit-learn
joblib
Flask-PyMongor